requirements.txt
amqp==2.3.2
billiard==3.5.0.4
celery==4.1.0
certifi==2018.8.13
chardet==3.0.4
click==6.7
Django==1.11.15
Flask==0.12.4
Flask-Uploads==0.1.3
idna==2.7
image==1.5.24
itsdangerous==0.24
Jinja2==2.10
kombu==4.2.1
MarkupSafe==1.0
numpy==1.15.1
opencv-python==3.4.2.17
Pillow==5.2.0
Pillow-SIMD==5.1.1.post0
pytz==2018.5
requests==2.19.1
urllib3==1.23
vine==1.1.4
Werkzeug==0.14.1
